educaplan/requirements.txt
ttkbootstrap==1.10.1
mysql-connector-python==8.0.33

